Use 'mon restart' instead of systemctl
On a few systems we have experienced problems with LMD not starting, due
to systemctl failing to 'restart' Naemon. The issue with systemctl not
being able to restart naemon, stops if one do a stop/start cycle
instead.

The mon restart script, does indeed use a stop/start cycle, and
therefore this fix uses the mon script instead as a workaround.

This fixes: MON-11574

Signed-off-by: Jacob Hansen <jhansen@op5.com>