Run LMD as monitor user instead of root
This commit changes the service / init files to run the LMD process as
the monitor user instead of the root user.

This change also included not changing the livestatus.cfg file and
restarting Naemon on everytime LMD starts, as restarting the service
requires root. LMD does not require Naemon to restart to work, and the
livestatus.cfg file is already edited on every update/install so it is not
required to do it on every startup of the LMD process.

This fixes: MON-11466

Signed-off-by: Jacob Hansen <jhansen@op5.com>