Explicitly set the group for the service
With this commit we explicitly set the group for the LMD systemd
service. In most cases this should not be required, however a customer
report mentioned that the groupright on the socket was set to
monitor:monitor making apache unable to write to the socket. This should
fix that case.

Signed-off-by: Jacob Hansen <jhansen@op5.com>