Add nonsense comment to specfile
Unfortunatley the master and maint/8.x branches both have the same HEAD
commit. This is a problem when we have tag:ed both repos for release.
Depending on which tag that was pushed latest either Monitor 9 (master)
or Monitor 8 (maint/8.x) will get a LMD package with wrong VERSION set.

This nonsene comment to the master branch will ensure that no erreneous
package will be released to Monitor 9.

Signed-off-by: Daniel Nilsson <dnilsson@itrsgroup.com>